Update Hdiode_code_tree.js
I took this file as it was residing on Vortex on 11/22/13, and added functionality to buttons on xml editor tree, to highlight changes in code mirror, and began reconstruction of non-XML editor elements to follow an object-oriented structure.  I am aiming to be able to generate multiple xml editors per page.